fix: security improvements
docs: README.md
chore: typescript types
chore: benchmark
chore: example app

+++ example/app.js
@@ -6,18 +6,38 @@ import timingSafeEqual from 'compare-timing-safe'
 
 import { Challenge, Verifier } from '../src/index.js'
 import { expressify } from './expressify.js'
-import { homePage, loginPage } from './page.js'
+import { homePage, userPage, loginPage } from './page.js'
 
 const __dirname = fileURLToPath(new URL('.', import.meta.url)).replace(/\/$/, '')
 
-const INITIAL_COMPLEXITY = 8
+const INITIAL_COMPLEXITY = 11
+const MAX_COMPLEXITY = 21 // ~120 avg secs to calculate the nonce
+const SECRET = 'change me, if you can'
 
+// prevent replay attacks - consider using redis with ttl
 const challengeStore = new Map()
+// increase complexity on every bad login attempt
 const userBadLoginStore = new Map()
-const userStore = new Map([['alice', 'alice']])
+// username password store
+const userStore = new Map([['alice', 'alice'], ['bob', 'bob']])
 
-const challenger = new Challenge('this-is-secret')
+// ----
+
+const challenger = new Challenge(SECRET)
 const verifier = new Verifier(challenger)
+
+// ---- helpers
+
+const calcComplexity = (badlogins = 0) =>
+  Math.min(MAX_COMPLEXITY, INITIAL_COMPLEXITY + badlogins)
+
+const getNstoreChallenge = async (complexity) => {
+  const challenge = await challenger.create(complexity)
+  challengeStore.set(challenge, complexity)
+  return challenge
+}
+
+// ----
 
 const app = express()
 
@@ -32,16 +52,25 @@ app.use('/js',
 
 app.get('/',
   (req, res) => {
-    res.type('html').end(homePage({ title: 'Home' }))
+    const { username } = req.cookies || {}
+    const page = username
+      ? userPage({ username })
+      : homePage({ title: 'Home', username })
+    res.type('html').end(page)
+  }
+)
+
+app.get('/logout',
+  (req, res) => {
+    res.clearCookie('username')
+    res.redirect('/')
   }
 )
 
 app.get('/login', expressify(
   async (req, res) => {
     const complexity = INITIAL_COMPLEXITY
-    const challenge = await challenger.create(complexity)
-    challengeStore.set(challenge, complexity)
-
+    const challenge = await getNstoreChallenge(complexity)
     res.type('html').end(loginPage({ challenge }))
   }
 ))
@@ -50,35 +79,41 @@ app.post('/login', expressify(
   bodyParser.urlencoded({ extended: false }),
   async (req, res) => {
     const { challenge, nonce, username, password } = req.body
+    let badlogins = 0
 
     try {
       if (!challengeStore.has(challenge)) {
-        throw new Error('Login failed! (No challenge in store)')
-      }
-      if (!await verifier.verify(challenge, nonce)) {
-        throw new Error('Login failed! (Bad nonce)')
-      }
-      if (!timingSafeEqual(userStore.get(username) || '', password || '')) {
-        throw new Error('Login failed! (Bad password)')
+        throw new Error('Login failed! (Replay)')
       }
       challengeStore.delete(challenge)
+
+      // calculate the complexity from previous bad login attempts for that user
+      // prevents using another simpler challenge from GET /login
+      badlogins = (userBadLoginStore.get(username) || 0)
+      const expectedComplexity = calcComplexity(badlogins)
+
+      if (!await verifier.verify(challenge, nonce, expectedComplexity)) {
+        throw new Error('Login failed! (Bad nonce)')
+      }
+      if (!username || !password || !timingSafeEqual(userStore.get(username) || '', password || '')) {
+        throw new Error('Login failed! (Bad password)')
+      }
       userBadLoginStore.delete(username)
+      res.cookie('username', username, { htmlOnly: true })
       res.redirect('/')
       return
     } catch (e) {
       console.error(e)
 
-      const badlogins = userBadLoginStore.get(username) || 1
-      userBadLoginStore.set(username, badlogins + 1)
+      badlogins += 1
+      userBadLoginStore.set(username, badlogins)
 
-      const complexity = Math.min(22, INITIAL_COMPLEXITY + badlogins)
-
+      const complexity = calcComplexity(badlogins)
       const error = e.message + `\n${badlogins}x bad login attempts.\nComplexity is ${complexity}`
 
-      const nextChallenge = await challenger.create(complexity)
-      challengeStore.set(nextChallenge, complexity)
-
-      res.type('html').end(loginPage({ challenge: nextChallenge, error, username }))
+      const nextChallenge = await getNstoreChallenge(complexity)
+      const simplechallenge = await getNstoreChallenge(INITIAL_COMPLEXITY)
+      res.type('html').end(loginPage({ challenge: nextChallenge, error, username, simplechallenge }))
     }
   }
 ))
